requirements.txt
diffusers>=0.37.1
transformers>=4.57.0
accelerate>=1.1.0
gradio>=5.0.0
Pillow>=10.0.0
safetensors>=0.4.5
sentencepiece>=0.2.0
numpy>=1.26.0
imageio>=2.36.0
imageio-ffmpeg>=0.5.1
ftfy>=6.3.1
